kernel: fix multi-page lending
Due to several errors, multi-page lending did not work properly.

To begin with, if a page was not allocated then it would not be able to
be sent. That is, if you allocated 8192 bytes but only ever accessed the
first 4096 bytes, then an error would be encountered when lending the
unaccessed upper page.

Furthermore, due to a striding error in accessing pointers, half-pages
were treated as full pages.

+++ kernel/src/server.rs
@@ -441,7 +441,7 @@ impl Server {
                     let client_tid = client_tid as _;
                     // Return the memory to the calling process
                     ss.return_memory(
-                        server_addr as *mut u8,
+                        server_addr as *mut usize,
                         client_pid,
                         client_tid,
                         client_addr as _,
